Add the widefocus layout
This layout puts the first window in the stack in a fixed region in the
center of the screen according to m->mfact. The next nmaster windows in
the stack are placed to the left, and the remaining stack is placed to
the right. The width ratio between the left and rigth stack is
determined by m->ffact